rf debug fs. rf module naming convention. feat: SS
