+++ src/components/dh_mappers/dh_m000/m000.rs
@@ -1,8 +1,8 @@
-use crate::components::dh_mappers::mapper::types::Mapper;
+use crate::components::dh_mappers::dh_mapper::mapper::Mapper;
 
 // #[derive(Debug, Clone, Copy)]
 // pub struct M000(MapperData); // test if i can just alias MapperData
 #[derive(Default)]
-pub struct MapsM000;
+pub struct MapOfM000;
 
-pub type M000 = Mapper<MapsM000>;
+pub type M000 = Mapper<MapOfM000>;
