+++ src/components/dh_mappers/dh_m000/mod.rs
@@ -1,2 +1,2 @@
 pub mod impls;
-pub mod types;
+pub mod m000;
