+++ src/components/dh_mappers/dh_mapper/impls/default.rs
@@ -2,7 +2,7 @@
 
 use std::marker::PhantomData;
 
-use crate::components::{self, dh_mappers::mapper::types::Mapper};
+use crate::components::{self, dh_mappers::dh_mapper::mapper::Mapper};
 
 impl<MapsType> Default for Mapper<MapsType>
 where
@@ -10,7 +10,7 @@ where
 {
     fn default() -> Self {
         // TODO: will cause some recursion
-        components::dh_mappers::mapper::types::Mapper {
+        components::dh_mappers::dh_mapper::mapper::Mapper {
             _marker: PhantomData,
             chr_bank: 0,
             prg_bank: 0,
