--- src/components/dh_mappers/mapper/mod.rs
@@ -1,2 +1,0 @@
-pub mod impls;
-pub mod types;
