+++ src/components/dh_mappers/mod.rs
@@ -1,3 +1,3 @@
-pub mod m000;
-pub mod mapper;
+pub mod dh_m000;
+pub mod dh_mapper;
 pub mod traits;
