+++ src/debug/types/dh_debuggee_message.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, PartialEq)]
+pub enum DebuggeeMessage {
+    Start,
+    KeyPressed(iced::keyboard::Key),
+    SyncHeader(iced::widget::scrollable::AbsoluteOffset),
+    End,
+}
