+++ src/debug/types/utilities.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone)]
+pub struct Utilities {
+    pub table_header_id: iced::widget::scrollable::Id,
+    pub table_body_id: iced::widget::scrollable::Id,
+    pub table_footer_id: iced::widget::scrollable::Id,
+}
