+++ src/debug/widgets/mod.rs
@@ -1,0 +1,3 @@
+pub mod cpu_monitor_view;
+pub mod grid_cell;
+pub mod ram_widgets;
