+++ src/debug/widgets/ram_widgets/mod.rs
@@ -1,0 +1,2 @@
+pub mod read_hits;
+pub mod write_hits;
